The project has been started in the monty of April 2018 and ended in April 2018
Established in 1993, today Silk land is a leading supplier of furnishing fabrics in the Middle East.Located in Dubai, Silkland has a global reach with offices in over 10 countries and is growing. Its vision and mission is to surpass all expectations by consistently offering well-merchandised collection of the most innovative and exclusive fabric designs to our customers which would help them create spaces for their dreams.